Update and clarify `dig` function documention.
Currently, the dig function's documentation doesn't realllllly explain
what specifically what the function does, or what arguments are
expected, unless you're already familiar with the general concept of a
dig function.

  * Clarified that dig expects to be called on a hash or array, as
opposed to the generic term "structure" which, while accurate, is less
informative to the inexperienced.
  * Quoted all bare strings being passed to the dig function.
    - https://docs.puppet.com/puppet/5.0/style_guide.html#quoting
  * Explain the difference between using dig and standard indexing, as
part of the example text. While this is sort of explained earlier, it's
far from clear without explaining it alongside an example case.